refactor: interface 파일 이름이 예약어와 겹치던 문제 해결
