+++ src/interface/LocationInterface.ts
